backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pdfplumber==0.11.4
PyMuPDF==1.24.14
anthropic==0.40.0
sqlalchemy==2.0.36
openpyxl==3.1.5
python-multipart==0.0.17
python-dotenv==1.0.1
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
